test for DeleteUSer
